+++ src/common/database/mapper/dbc/PgDbcCharStartOutFit.hpp
@@ -18,7 +18,7 @@ struct DbcCharStartOutfit {
     uint8_t RaceID;
     uint8_t ClassID;
     uint8_t SexID;
-    int16_t OutfitID;
+    uint8_t OutfitID;
 
     int32_t ItemID[24];
     int32_t DisplayItemID[24];
@@ -30,11 +30,12 @@ struct PgRowMapper<DbcCharStartOutfit> {
     static DbcCharStartOutfit map(const pqxx::row& r) {
         DbcCharStartOutfit row;
 
-        row.ID = static_cast<uint32_t>(r["id"].as<int32_t>());
-        row.RaceID = static_cast<uint8_t>(r["raceid"].as<int16_t>());
-        row.ClassID = static_cast<uint8_t>(r["classid"].as<int16_t>());
-        row.SexID = static_cast<uint8_t>(r["sexid"].as<int16_t>());
-        row.OutfitID = r["outfitid"].as<int16_t>();
+        // Прямое чтение числовых полей
+        row.ID       = r["id"].as<uint32_t>();
+        row.RaceID   = static_cast<uint8_t>(r["raceid"].as<int>());
+        row.ClassID  = static_cast<uint8_t>(r["classid"].as<int>());
+        row.SexID    = static_cast<uint8_t>(r["sexid"].as<int>());
+        row.OutfitID = static_cast<uint8_t>(r["outfitid"].as<int>());
 
         // itemid_1..24
         for (int i = 0; i < 24; i++) {
