+++ src/common/database/mapper/dbc/PgDbcSkillRaceClassInfo.hpp
@@ -29,14 +29,14 @@ struct PgRowMapper<DbcSkillRaceClassInfo> {
     static DbcSkillRaceClassInfo map(const pqxx::row& r) {
         DbcSkillRaceClassInfo row;
 
-        row.ID            = static_cast<uint32_t>(r["id"].as<int32_t>());
-        row.SkillID       = static_cast<uint32_t>(r["skillid"].as<int32_t>());
-        row.RaceMask      = static_cast<uint32_t>(r["racemask"].as<int32_t>());
-        row.ClassMask     = static_cast<uint32_t>(r["classmask"].as<int32_t>());
-        row.Flags         = static_cast<uint32_t>(r["flags"].as<int32_t>());
-        row.MinLevel      = static_cast<uint32_t>(r["minlevel"].as<int32_t>());
-        row.SkillTierID   = static_cast<uint32_t>(r["skilltierid"].as<int32_t>());
-        row.SkillCostIndex= static_cast<uint32_t>(r["skillcostindex"].as<int32_t>());
+        row.ID            = r["id"].as<uint32_t>();
+        row.SkillID       = r["skillid"].as<uint32_t>();
+        row.RaceMask      = r["racemask"].as<uint32_t>();
+        row.ClassMask     = r["classmask"].as<uint32_t>();
+        row.Flags         = r["flags"].as<uint32_t>();
+        row.MinLevel      = r["minlevel"].as<uint32_t>();
+        row.SkillTierID   = r["skilltierid"].as<uint32_t>();
+        row.SkillCostIndex= r["skillcostindex"].as<uint32_t>();
 
         return row;
     }
